django-rest/escola/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-filter==24.1
djangorestframework==3.15.0
drf-yasg==1.21.10
Faker==24.3.0
inflection==0.5.1
Markdown==3.6
packaging==25.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.2.0
validate-docbr==1.10.0
